vpc-sample/run-egress/requirements.txt
Flask==2.2.3
gunicorn==20.1.0
google-auth==2.22.0
requests==2.31.0
